refac(91026): Extrai task processa_carga_async
