+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -1,3 +1,4 @@
 from .tasks_legado import *
 
 from .concluir_prestacao_de_contas import concluir_prestacao_de_contas_async
+from .processa_carga import processa_carga_async
