+++ sme_ptrf_apps/core/tasks/processa_carga.py
@@ -1,0 +1,23 @@
+import logging
+
+from celery import shared_task
+
+from sme_ptrf_apps.core.models import (
+    Arquivo,
+)
+
+logger = logging.getLogger(__name__)
+
+@shared_task(
+    retry_backoff=2,
+    retry_kwargs={'max_retries': 8},)
+def processa_carga_async(arquivo_uuid):
+    logger.info("Iniciando task processa_carga_async")
+    from sme_ptrf_apps.core.services import processa_carga
+    logger.info("Processando arquivo %s", arquivo_uuid)
+    arquivo = Arquivo.objects.filter(uuid=arquivo_uuid).first()
+    if not arquivo:
+        logger.info("Arquivo não encontrado %s", arquivo_uuid)
+    else:
+        logger.info("Arquivo encontrado %s", arquivo_uuid)
+    processa_carga(arquivo)
